"feat: 게시글 목록 조회 API 구현 (인피니티 스크롤)
- QueryDSL 설정 및 커서 기반 페이징
- N+1 문제 해결 (JOIN 활용)
- PostRepositoryCustom + Impl 추가
- PostListResponseDto 추가"

+++ src/main/java/ktb/cloud_james/community/entity/PostStats.java
@@ -44,10 +44,10 @@ public class PostStats {
     private Long viewCount = 0L;
 
     @Column(name = "like_count", nullable = false)
-    private Integer likeCount = 0;
+    private Long likeCount = 0L;
 
     @Column(name = "comment_count", nullable = false)
-    private Integer commentCount = 0;
+    private Long commentCount = 0L;
 
     /**
      * Builder 없이 생성자만 제공
